[Feature:Developer] Allow customization of VM resources (#11922)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Currently whenever `vagrant up` is run the VM is created with 2G of ram
and 2 cpus.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
This adds two new environment variables that can be used to configure
the values for the vm. An example of both of them used is
`VM_MEMORY=4096 VM_CPUS=4 vagrant up`

### What steps should a reviewer take to reproduce or test the bug or
new feature?
1. Run `vagrant up` and observe the memory and cpus allocated to your
vm.
2. Run `vagrant halt && vagrant up` with `VM_CPUS` and/or `VM_MEMORY`
set and verify the correct value is used.

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

documentation added in PR:
https://github.com/Submitty/submitty.github.io/pull/695

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->