[Bugfix:System] Fix add user script (#11925)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
The adduser.py script in /sbin was updated with sqlalchemy update, and a
variable was passed over when updated, so the script fails to run.
### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
That variable has been updated, now the script runs as expected. 
### What steps should a reviewer take to reproduce or test the bug or
new feature?
Try and run the script on main, then switch to the branch and try and
run it there.
### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
This was found when testing for
https://github.com/Submitty/Submitty/pull/11530
### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->