[Refactor:TAGrading] Move student info panel to vue (#11889)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
This moves the second tagrading panel to vue.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Student info is rendered with vue and the twig + separate scripts are
removed.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
1. Start TAGrading
2. View the student info panel
3. Check that functionality should be unchanged.

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
The student info panel and version choice box already are thoroughly
tested.

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->