[Bugfix:Developer] Fix eslint config on Windows (#11976)
### Why is this Change Important & Necessary?

Right now it's impossible for me to run eslint as a windows user. I'm
not sure what the situation is for other developers.

### What is the New Behavior?

running npx eslint in the site folder should correctly run the
Javascript linter.

### What steps should a reviewer take to reproduce or test the bug or
new feature?

Try to run eslint outside for your VM in the site folder on main. Switch
to this branch, npm install in site, and run `npx eslint` or `npm run
eslint`

### Automated Testing & Documentation

No testing or docs update is needed.

### Other information
N/A