[Testing:Developer] Rust Autograding (#11926)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Part of #11641.
Currently Rust Autograding is not working due to syscalls. This is the
fix and the testing of the rust autograding.
Base branch is #11915

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Rust autograding is being tested. In addition, the various syscalls for
rust is added to both of the submitty rust autograding config. In order
to get compilation working, I had to allow all syscalls even though I
ran the detecting syscall function and allowed all syscalls there.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Make sure that the test is not flaky

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
Unit tests are not possible due to docker

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>