[Refactor:System] Fix randint types in sample course data (#11975)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Currently we pass floats to `random.randint` when setting up sample
courses, and this currently is supported but will not be when our python
version upgrades along with ubuntu 24.04.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Floats are casted to ints before being passed to `random.randint`

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Verify that the sample courses are created and function normally.

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
Almost all of our tests are based on the sample data, so this change is
thoroughly covered.

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->