[Bugfix:InstructorUI] Fix decoding of eg_thread_ids (#11977)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
From #11702,
On production, we have gradeables that have already been created. When
the gradeable gets created, we `json_encode('{}')` as the default value.
This means that when it gets inserted into the database, instead of `{}`
being inserted, we actually insert `"{}"` instead, which is a string.
This means that the new code, which decodes it once, will not work as it
will be decoded to `"{}"` instead of `[]`, which throws of frog robot.
We can either:

1. Decoding it twice would allow for us to get `[]` instead of string
`"{}"`.
2. Make a database migration for this situation where we specifically
patch this one instance

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Make a database migration for this situation where we specifically patch
this one instance. I feel like this is better and the other way is a
little hacky

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Make a gradeable before #11702. Move to after #11702, edit gradeable,
see error

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
This could not have been prevented that easily with e2e test. The
initial state would have to be from the previous course creations, and
then switching to the new version. We could do this to prevent this
error, but it would be very hard.

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->