[Bugfix:InstructorUI] Fix checkpoint dates (#11983)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
#11539 added this line. This line does not work for checkpoint
gradeables as `gradeable.js` does not get included. This would cause all
the dates to fail as the event listeners below it are in control of the
dates

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
We should only load bulk uploads if we have an electronic gradeable

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Visit the checkpoint gradeable edit page. Notice error in console. Go to
dates, notice that you cannot change any as a refresh will clear it

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
#11982

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->