[Bugfix:TAGrading] Fix Saving rubric to navigate (#11984)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Currently whenever you click next student button and have a component
open, a popup displays that says that changes will not be saved. This
happens because in Vue, `GRADED_GRADEABLE` does not exist in vue but it
does exist in the typescript. This causes an annoyance which is also
simply not true

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Checks for graded gradeable before executing a function that requires it
to not be null

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Open a component. Click next. See error. Change to this PR, does not see
error.

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
This is not tested by e2e tests.

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->