[Refactor:Developer] Flatten Autograding Testing Suite (#11927)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Looking at #11926, every time we want to submit to a new gradeable we
need to go a layer deeper. This majorly impacts readability and makes it
harder for people to make new tests
Base branch is #11926 

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Makes it easier to submit autograding tests. Uses `.then` instead of
chainging .then. Since most of the autograding commands are the same,
ive also refactorted it such that we only have two commands, one to
submit the files and one to check the results

### What steps should a reviewer take to reproduce or test the bug or
new feature?

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>