[Refactor:System] Make Terms Doctrine (#11787)
### Why is this Change Important & Necessary?
Refactors terms to use Doctrine.

### What is the New Behavior?
Terms are now a doctrine entity, this is a rarely used part of the
system, however it will allow for converting courses to doctrine easier,
as most of the 'select course...' queries require selecting from term
database.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
If you create a new term as 'superuser' in the 'create new course' page,
and it successfully passes, and you can view it in the 'select term'
dropdown, this has worked.

You can also verify in the database. 

### Automated Testing & Documentation
A unit test has been created, but more tests may be added later if the
functionality increases.
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->

---------

Co-authored-by: Alex Lavallee <alex@lavallee.one>