[Feature:TAGrading] Withdrawn student grading toggle (#11879)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
closes #11874 
We are automatically including the withdrawn student in the grading
statistics, leading to incomplete ta results. More could be read in the
issue above. We want a toggle that automatically hides all the withdrawn
students so we can grade all the "active" students.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Synced the two cookies to `include_withdrawn_students`  
Made the statistics correct based on if we are hiding or showing
withdrawn students.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Make sure that teams and user gradeables both work. Make sure the code
is consistent and that each stat changes with cookie changes. Make sure
that team gradeable stats don't change.

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
[#11763](https://github.com/Submitty/Submitty/pull/11882) starts with
adding withdrawn students. The tests will be added after that.

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->