[Testing:Developer] Explicit fail on cypress unmet condition (#12004)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
DockerUI has been failing rarely now. I cannot debug the issue
successfully, but I did notice that there was some discrency in the
tests. This is because the dockerui checks the docker status, but that
could fail in the previous test. When we use `cy.waitAndReloadUntil`, we
do not really fail if the condition fails to be met. This causes the CI
to fail in general.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
This explicitly fails if our condition is never met. Otherwise, the test
runs as usual.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
```js
describe('123', () => {
    it('Test Case', () => {
        cy.login();
        cy.visit('/admin/docker');
        // eslint-disable-next-line no-restricted-syntax
        cy.waitAndReloadUntil(() => {
            return cy.get('[data-testid="docker-status"]')
                .invoke('text')
                .then((text) => {
                    return text.includes('Text does not exist');
                });
        }, 10000, 500);
    });
});
```
This grabs something that does not exist. Originally the behavior would
be that the test passed. With the newbehavior, an error is correctly
thrown

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->
The dockerui might need other work