[Feature:TAGrading] Withdrawn Student Banner (#12005)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
There is currently no way for a TA to distinguish who is a withdrawn
student or not. If they navigate using the navigating button, they might
accidentally grade a withdrawn student, who doesn't actually need
grading.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
This PR aims to add banners to distinguish the withdrawn status. In
addition, this slightly fixes a bug where a overridden student submits
to a gradeable and the text not appearing on the rubric panel.

<img width="1055" height="692" alt="image"
src="https://github.com/user-attachments/assets/1f49e419-ffda-4c18-b798-7ccd82f54317"
/>

### What steps should a reviewer take to reproduce or test the bug or
new feature?

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
#11882 will add withdrawn students, then it would be easier to test

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->