[Refactor:Forum] Refactor Chatroom Row (#11863)
### Why is this Change Important & Necessary?
Before, the way we handled chatroom toggles worked well for the standard
usage of chatrooms, with an instructor broadcasting chatroom enable and
a student receiving chatroom enabling, which then triggers a row insert.
This had the negative side effect of being difficult to update the
chatroom row style because we had to make changes in two places.
Additionally, since the templates were separate the output was
inconsistent depending on what role you had, causing buggy looking
chatroom rows.

### What is the New Behavior?
Now, a public template named ChatroomRow is added, which is rendered
using twig.js. When a chatroom is updated it will delete the row and
recreate it using the twig template.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Open two chatroom list pages logged in as instructor, create and start a
chatroom session. It should look buggy/not add the chatroom at all. On
this branch it should function as expected between two instructor
windows.

### Automated Testing & Documentation
This is now tested through #11818, and I added websocket enable and
disable tests, meaning that the websocket features of live chat would be
fully tested up to now.

### Other information
Potential TODO use vue instead of Twig.js

---------

Co-authored-by: Justin Manion <jmanion32@gmail.com>