Removed Index, Commented Out Mark All, Added Reactive Updating
Index no longer needed (for now) since we are not doing the mark all as seen option. I commented out mark all as seen but did not delete so that it is easy to bring back in the future if necessary. Made the notifications array reactive so we can apply updates to its seen/unseen status without having to reload the page.