[Bugfix:Developer] Optional Notebook Elements Fix (#12016)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
If an instructor uploads a notebook that has an image that has a defined
height or width, then an image that does have a defined heigth or width,
it will be fine on production but when we get to a developer
environment, there will a lot of errors because of strict_type.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
This PR aims to check if optional elements exists and then apply them if
possible

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Redacted psoft to see error

[Archive.zip](https://github.com/user-attachments/files/22053182/Archive.zip)

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
N/A

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->

Co-authored-by: Viane Matsibekker <04vmatsibekker@gmail.com>