[Bugfix:Submission] Fix vue-id collision (#12010)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
@lavalleeale and I knew that this day would come, where we would have so
many vue elements on the screen that the id's conflict with each other
and make a very weird site. Therefore, we need to stop using random ids
and instead be asserting that it is per script. There are many mistakes
in teh gradeable itself, but the part I can show demonstrates it
perfectly.
<img width="1293" height="442" alt="image"
src="https://github.com/user-attachments/assets/9e283908-95f3-4295-bdce-af2800402bfc"
/>
Write your java code is a markdown box that is a vue element. On this
page, it seems like it replaces the too many queries banner, the
sidebar, and some other things that I did not want to pinpoint. I can
confirm this collision as opening it up and checking for the specific
ID, there are going to multiple,

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Instead of generating a random ids, we will put the script tag in the
div. This also works for removing the script tag as the vue element
renders on top on it.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
I don't know how you can test this reliable, maybe have a large
markdown/notebook gradeable.

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>