[Feature:Autograding] Add regrade_by to autograding history (#12006)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Whenever we click regrade, we have no idea who the submission has been
graded by. If we want to include students regrading, we should have an
idea of who regraded the submission. Therefore, we should add this in
the queue file, which adds it to the history file.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Adds the user who regrades into the queue and history file

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Make sure that autograding does not break. I have tested team, vcs, and
notebook submissions. I have also verified that the history with
regrade_by is written to the file.

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->
Unit tests are included. E2E tests would still have to be written.

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->